cmd/go: enhance the tests for Mercurial vcs.* stamped data
The timestamp and revision values can be wrong, so start by applying a fixed
configuration that will keep the values stable in future test runs, and then
test for those values.  Also, add a test for a configuration that wrongly
interprets `hg status` output as meaning there were modifications.  I left the
status call in the test to show an example of optional output, and tested for it
so the test scenario doesn't silently regress.